refactor: migrate to TS
