--- src/Test.css
@@ -1,136 +1,0 @@
-.app {
-  background: #0097c1;
-  min-width: 80%;
-  min-height: 200px;
-  height: min-content;
-  border-radius: 5px;
-  padding: 20px;
-  box-shadow: 10px 10px 42px 0px black;
-  display: flex;
-  justify-content: space-evenly;
-  font-size: 1.2rem;
-  font-weight: 600;
-  color: white;
-}
-
-.score-section {
-  display: flex;
-  font-size: 24px;
-  align-items: center;
-}
-
-/* QUESTION/TIMER/LEFT SECTION */
-.question-section {
-  width: 100%;
-  position: relative;
-}
-
-.question-count {
-  margin-bottom: 20px;
-}
-
-.question-count span {
-  font-size: 28px;
-}
-
-.question-text {
-  margin-bottom: 12px;
-}
-
-/* ANSWERS/RIGHT SECTION */
-.answer-section {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-}
-
-.test-button {
-  min-width: 80px;
-  min-height: 35px;
-  line-height: 35px;
-  font-size: 18px;
-  font-family: sans-serif;
-  text-decoration: none;
-  background-color: transparent;
-  letter-spacing: 2px;
-  text-align: center;
-  color: #fff;
-  border: 2px solid #fff;
-  transition: all 0.35s;
-}
-
-.test-text {
-  position: relative;
-  z-index: 2;
-}
-
-.test-button:after {
-  position: absolute;
-  content: "";
-  top: 0;
-  left: 0;
-  width: 0;
-  height: 100%;
-  background: #75f8e2;
-  transition: all 0.35s;
-}
-
-.test-button:hover {
-  background-color: transparent;
-  border: 2px solid #fff;
-}
-
-.test-button:hover:after {
-  width: 100%;
-}
-
-/* 
-.quiz-button {
-  min-width: 80px;
-  min-height: 35px;
-  line-height: 35px;
-  font-size: 18px;
-  font-family: sans-serif;
-  text-decoration: none;
-  background-color: transparent;
-  letter-spacing: 2px;
-  text-align: center;
-  transition: all 0.35s;
-width: 100%;
-  font-size: 1.1rem;
-  font-weight: 600;
-  color: white;
-  background: #61dafb;
-  border-radius: 5px;
-  display: flex;
-  padding: 5px;
-  justify-content: flex-start;
-  align-items: center;
-  border: 3px solid #61dafb;
-  cursor: pointer;
-}
-
-.quiz-button::selection {
-  background: transparent;
-}
-
-.quiz-button:after {
-  position: absolute;
-  content: "";
-  top: 0;
-  left: 0;
-  width: 0;
-  height: 100%;
-  background: #ff003b;
-  transition: all 0.35s;
-}
-
-.quiz-button:hover {
-  color: #fff;
-  background-color: transparent;
-  background-color: #75f8e2;
-  border: 3px solid #75f8e2;
-}
-
-*/
