--- src/TestPage.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const TestPage = () => {
-  return <>Test</>;
-};
-
-export default TestPage;
