+++ src/components/Level/LevelThree.tsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { Title } from "../../stories/headers";
 import { PrimaryButton, SkipButton } from "../../stories/buttons/";
-import { useLockNoPrompt, useUnlockNoPrompt } from "../../utils/utils";
+import { useLockNoPrompt, useUnlockNoPrompt } from "../../utils/lockNavigation";
 import { QuestionForm } from "../../stories/forms/";
 import { PrivateRoutes } from "../../PrivateRoutes";
 import { modes } from "../../flags";
