--- src/flags.js
@@ -1,3 +1,0 @@
-export const featFlags = { test: true, pd: false };
-
-export const modes = { promptMode: false, skipMode: true, quizMode: true };
