+++ src/flags.ts
@@ -1,0 +1,3 @@
+export const featFlags = { test: true, pd: false };
+
+export const modes = { promptMode: false, skipMode: true, quizMode: false };
