+++ src/utils/cryptoLib.ts
@@ -1,6 +1,7 @@
-export const crypt = (salt, text) => {
-  const textToChars = (text) => text.split("").map((c) => c.charCodeAt(0));
-  const byteHex = (n) => ("0" + Number(n).toString(16)).substr(-2);
+export const crypt = (salt: string, text: string) => {
+  const textToChars = (text: string) =>
+    text.split("").map((c) => c.charCodeAt(0));
+  const byteHex = (n: number) => ("0" + Number(n).toString(16)).substr(-2);
   const applySaltToChar = (code) =>
     textToChars(salt).reduce((a, b) => a ^ b, code);
 
@@ -12,15 +13,16 @@ export const crypt = (salt, text) => {
     .join("");
 };
 
-export const decrypt = (salt, encoded) => {
-  const textToChars = (text) => text.split("").map((c) => c.charCodeAt(0));
+export const decrypt = (salt: string, encoded: any) => {
+  const textToChars = (text: string) =>
+    text.split("").map((c) => c.charCodeAt(0));
   const applySaltToChar = (code) =>
     textToChars(salt).reduce((a, b) => a ^ b, code);
   return encoded
     .match(/.{1,2}/g)
-    .map((hex) => parseInt(hex, 16))
+    .map((hex: string) => parseInt(hex, 16))
     .map(applySaltToChar)
-    .map((charCode) => String.fromCharCode(charCode))
+    .map((charCode: number) => String.fromCharCode(charCode))
     .join("");
 };
