+++ src/utils/useLocalStorage.ts
@@ -1,7 +1,7 @@
 import { useState } from "react";
 
 // Hook
-export function useLocalStorage(key, initialValue) {
+export function useLocalStorage(key: string, initialValue: string) {
   // State to store our value
   // Pass initial state function to useState so logic is only executed once
   const [storedValue, setStoredValue] = useState(() => {
@@ -21,7 +21,7 @@ export function useLocalStorage(key, initialValue) {
   });
   // Return a wrapped version of useState's setter function that ...
   // ... persists the new value to localStorage.
-  const setValue = (value) => {
+  const setValue = (value: any) => {
     try {
       // Allow value to be a function so we have same API as useState
       const valueToStore =
